feat(comment): using the article controller/paths
