+++ src/articles/articles.module.ts
@@ -3,9 +3,10 @@ import { ArticlesService } from './articles.service';
 import { ArticlesController } from './articles.controller';
 import { RelationalArticlePersistenceModule } from './infrastructure/persistence/relational/relational-persistence.module';
 import { UsersModule } from '../users/users.module';
+import { CommentsModule } from '../comments/comments.module';
 
 @Module({
-  imports: [RelationalArticlePersistenceModule, UsersModule],
+  imports: [RelationalArticlePersistenceModule, UsersModule, CommentsModule],
   controllers: [ArticlesController],
   providers: [ArticlesService],
   exports: [ArticlesService, RelationalArticlePersistenceModule],
