+++ src/comments/infrastructure/persistence/comment.repository.ts
@@ -1,5 +1,3 @@
-import { DeepPartial } from '../../../utils/types/deep-partial.type';
-import { NullableType } from '../../../utils/types/nullable.type';
 import { IPaginationOptions } from '../../../utils/types/pagination-options';
 import { Comment } from '../../domain/comment';
 
@@ -16,13 +14,6 @@ export abstract class CommentRepository {
   }: {
     paginationOptions: IPaginationOptions;
   }): Promise<Comment[]>;
-
-  abstract findById(id: Comment['id']): Promise<NullableType<Comment>>;
-
-  abstract update(
-    id: Comment['id'],
-    payload: DeepPartial<Comment>,
-  ): Promise<Comment | null>;
 
   abstract remove(id: Comment['id']): Promise<void>;
 }
