+++ src/comments/infrastructure/persistence/relational/repositories/comment.repository.ts
@@ -2,7 +2,6 @@ import { Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { CommentEntity } from '../entities/comment.entity';
-import { NullableType } from '../../../../../utils/types/nullable.type';
 import { Comment } from '../../../../domain/comment';
 import { CommentRepository } from '../../comment.repository';
 import { CommentMapper } from '../mappers/comment.mapper';
@@ -34,35 +33,6 @@ export class CommentRelationalRepository implements CommentRepository {
     });
 
     return entities.map((entity) => CommentMapper.toDomain(entity));
-  }
-
-  async findById(id: Comment['id']): Promise<NullableType<Comment>> {
-    const entity = await this.commentRepository.findOne({
-      where: { id },
-    });
-
-    return entity ? CommentMapper.toDomain(entity) : null;
-  }
-
-  async update(id: Comment['id'], payload: Partial<Comment>): Promise<Comment> {
-    const entity = await this.commentRepository.findOne({
-      where: { id },
-    });
-
-    if (!entity) {
-      throw new Error('Record not found');
-    }
-
-    const updatedEntity = await this.commentRepository.save(
-      this.commentRepository.create(
-        CommentMapper.toPersistence({
-          ...CommentMapper.toDomain(entity),
-          ...payload,
-        }),
-      ),
-    );
-
-    return CommentMapper.toDomain(updatedEntity);
   }
 
   async remove(id: Comment['id']): Promise<void> {
